continuous interaction on pairwise contrasts
Effect of continuous interaction on pairwise contrasts. When I have a model that includes the interaction between a categorical variable and a continuous one and I want to compare the effect of the continous var on the pairwise contrast. Here the hypothesis is that for an increase in the continuous var, the difference between the different levels of the categorical variable should be different --> grows or decreases